* Fixed: compiling errors under windows.

+++ GfxRaw/RawGraph.cpp
@@ -799,8 +799,8 @@ void CRawGraph::bltImage(int xDest, int yDest, int widthDest, int heightDest, Ra
     CRect rcDst = m_rcClipScaleMaped;
 
     // get the clip box to copy
-    rcDst.right = min(rcDst.right, xDest + widthDest);
-    rcDst.bottom = min(rcDst.bottom, yDest + heightDest);
+    rcDst.right = min((int)rcDst.right, xDest + widthDest);
+    rcDst.bottom = min((int)rcDst.bottom, yDest + heightDest);
 
     // determine the start position of Graph, image, mask.
     if (xDest >= rcDst.left) {
